# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: RickAndMorty
  version: 1.0.0
servers:
  - url: http://localhost:8876/api
...
paths:
...
  /episode:
    get:
      parameters:
        - name: ids
          in: query
          required: false
          schema:
            type: array
            items:
              type: number
            example: [1, 2]
        - name: page
          in: query
          required: false
          schema:
            type: number
            example: 1
        - name: name
          in: query
          required: false
          schema:
            type: string
            example: "Rick Sanchez"
        - name: code
          in: query
          required: false
          schema:
            type: string
            example: "S01E01"
      tags:
        - episodes
      summary: Get many episodes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedEpisodes"
        "400":
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - episodes
      summary: Create episode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEpisode"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEpisode"
        "400":
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /episode/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            example: 1
      tags:
        - episodes
      summary: Get a single episode
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEpisode"

    put:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            example: 1
      tags:
        - episodes
      summary: Update episode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEpisode"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEpisode"
        "400":
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            example: 1
      tags:
        - episodes
      summary: Change episode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEpisode"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEpisode"
        "400":
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            example: 1
      tags:
        - episodes
      summary: Delete episode
      responses:
        "200":
          description: OK
...
components:
  schemas:
    Error:
      type: object
      properties:
        name:
          type: string
          example: "Object(App\\DTO\\In\\Episode\\GetEpisodesDto).filters.code: This value is not valid."
    Paginated:
      type: object
      properties:
        count:
          type: integer
          example: 826
        pages:
          type: integer
          example: 42
        next:
          type: string
          nullable: true
          format: uri
          example: "http://localhost:8876/api/character/?page=3"
        prev:
          type: string
          nullable: true
          format: uri
          example: "http://localhost:8876/api/character/?page=1"
...
    GetEpisode:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: "Pilot"
        air_date:
          type: string
          example: "December 2, 2013"
        episode:
          type: string
          example: "S01E01"
        characters:
          type: array
          items:
            type: string
            format: uri
            example: "http://localhost:8876/api/character/1"
        url:
          type: string
          format: uri
          example: "http://localhost:8876/api/episode/1"
        created:
          type: string
          format: date-time
          example: "2017-11-04T18:48:46.250Z"
    PaginatedEpisodes:
      type: object
      properties:
        info:
          $ref: '#/components/schemas/Paginated'
        results:
          type: array
          items:
            $ref: '#/components/schemas/GetEpisode'
    PostEpisode:
      type: object
      properties:
        name:
          type: string
          example: "Pilot"
        air_date:
          type: string
          example: "December 2, 2013"
        episode:
          type: string
          example: "S01E01"
...